# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/campaigns/ignore": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Mark campaign as ignored on queue",
                "operationId": "34efde17c16882afa56249fc795ba073",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign marked as ignored on queue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "campaign": {
                                            "type": "object"
                                        },
                                        "queue_stats": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Campaign not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/campaigns/unignore": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Mark campaign as not ignored on queue",
                "operationId": "78208b80cbad0a863d3a69e456485ca7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign marked as not ignored on queue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "campaign": {
                                            "type": "object"
                                        },
                                        "queue_stats": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Campaign not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}